+++ pushnotifications/src/main/java/com/pusher/pushnotifications/PushNotificationsInstance.kt
@@ -165,11 +165,18 @@ class PushNotificationsInstance @JvmOverloads constructor(
     startHasBeenCalledThisSession = true
     val handleFcmToken = { fcmToken: String ->
       synchronized(deviceStateStore) {
-        if (deviceStateStore.startJobHasBeenEnqueued) {
-          serverSyncHandler.sendMessage(ServerSyncHandler.refreshToken(fcmToken))
-        } else {
-          serverSyncHandler.sendMessage(ServerSyncHandler.start(fcmToken, oldSDKDeviceStateStore.clientIds()))
-          deviceStateStore.startJobHasBeenEnqueued = true
+        if (startHasBeenCalledThisSession) {
+          if (deviceStateStore.startJobHasBeenEnqueued) {
+            serverSyncHandler.sendMessage(ServerSyncHandler.refreshToken(fcmToken))
+          } else {
+            serverSyncHandler.sendMessage(
+              ServerSyncHandler.start(
+                fcmToken,
+                oldSDKDeviceStateStore.clientIds()
+              )
+            )
+            deviceStateStore.startJobHasBeenEnqueued = true
+          }
         }
       }
 
@@ -361,7 +368,7 @@ class PushNotificationsInstance @JvmOverloads constructor(
    * @param callback callback used to indicate whether the user association process has succeeded
    */
   @JvmOverloads
-  fun setUserId(userId: String, tokenProvider: TokenProvider, callback: BeamsCallback<Void, PusherCallbackError> = NoopBeamsCallback()) {
+  fun setUserId(userId: String, tokenProvider: TokenProvider?, callback: BeamsCallback<Void, PusherCallbackError> = NoopBeamsCallback()) {
     if (tokenProvider == null) { // this can happen when using Java
       throw IllegalStateException("Token provider can't be null")
     }