+++ pushnotifications/src/main/java/com/pusher/pushnotifications/api/PushNotificationsAPI.kt
@@ -104,6 +104,11 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
   // Handles the JsonSyntaxException properly
   private fun safeExtractJsonError(possiblyJson: String): NOKResponse {
     return try {
+      // This if check is added to make it work as old Kotlin version.
+      // Need to be revisited in future
+      if (possiblyJson.trim() == ""){
+        throw IllegalStateException("Empty json string")
+      }
       gson.fromJson(possiblyJson, NOKResponse::class.java)
     } catch (jsonException: JsonSyntaxException) {
       log.w("Failed to parse json `$possiblyJson`", jsonException)
@@ -136,10 +141,10 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         // not throwing a `PushNotificationsAPIBadRequest` here so that it still retries this request.
         // it would make the code that calls this more complex to handle it.
         // this really shouldn't happen anyway
-        log.e("Critical error when registering a new device (error body: ${response?.errorBody()})")
+        log.e("Critical error when registering a new device (error body: ${response.errorBody()})")
       }
 
-      val responseBody = response?.body()
+      val responseBody = response.body()
       if (responseBody != null && response.code() in 200..299) {
         if (responseBody.initialInterestSet==null || responseBody.id==null){
           throw PushNotificationsAPIException("device id or initial interests is null")
@@ -149,7 +154,7 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
             initialInterests = responseBody.initialInterestSet)
       }
 
-      val responseErrorBody = response?.errorBody()
+      val responseErrorBody = response.errorBody()
       if (responseErrorBody != null) {
         val error = safeExtractJsonError(responseErrorBody.string())
         log.w("Failed to register device: $error")
@@ -171,12 +176,12 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to subscribe to interest: $error")
@@ -199,12 +204,12 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to unsubscribe from interest: $error")
@@ -227,12 +232,12 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to set subscriptions: $error")
@@ -255,12 +260,12 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to refresh FCM token: $error")
@@ -287,12 +292,12 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to set device metadata: $error")
@@ -316,30 +321,30 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         throw PushNotificationsAPIDeviceNotFound()
       }
       if (response.code() == 400) {
-        val reason = response?.errorBody()?.let { safeExtractJsonError(it.string()).description }
+        val reason = response.errorBody()?.let { safeExtractJsonError(it.string()).description }
         throw PushNotificationsAPIBadRequest(reason ?: "Unknown reason")
       }
       if (response.code() == 401 || response.code() == 403) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
-          throw PushNotificationsAPIBadJWT("${error?.error}: ${error?.description}")
+          throw PushNotificationsAPIBadJWT("${error.error}: ${error.description}")
         }
 
         throw PushNotificationsAPIBadJWT("Unknown reason")
       }
       if (response.code() == 422) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
-          throw PushNotificationsAPIUnprocessableEntity("${error?.error}: ${error?.description}")
+          throw PushNotificationsAPIUnprocessableEntity("${error.error}: ${error.description}")
         }
 
         throw PushNotificationsAPIUnprocessableEntity("Unknown reason")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to set user id: $error")
@@ -367,11 +372,11 @@ class PushNotificationsAPI(private val instanceId: String, overrideHostURL: Stri
         // not throwing a `PushNotificationsAPIBadRequest` here so that it still retries this request.
         // it would make the code that calls this more complex to handle it.
         // this really shouldn't happen anyway
-        log.e("Critical error when deleting a device (error body: ${response?.errorBody()})")
+        log.e("Critical error when deleting a device (error body: ${response.errorBody()})")
       }
 
       if (response.code() !in 200..299) {
-        val responseErrorBody = response?.errorBody()
+        val responseErrorBody = response.errorBody()
         if (responseErrorBody != null) {
           val error = safeExtractJsonError(responseErrorBody.string())
           log.w("Failed to delete device: $error")