+++ pushnotifications/src/main/java/com/pusher/pushnotifications/reporting/OpenNotificationActivity.kt
@@ -3,10 +3,9 @@ package com.pusher.pushnotifications.reporting
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import com.firebase.jobdispatcher.*
+import androidx.work.*
 import com.google.gson.Gson
 import com.google.gson.JsonSyntaxException
-import com.pusher.pushnotifications.internal.DeviceStateStore
 import com.pusher.pushnotifications.internal.InstanceDeviceStateStore
 import com.pusher.pushnotifications.logging.Logger
 import com.pusher.pushnotifications.reporting.api.OpenEvent
@@ -20,13 +19,13 @@ import com.pusher.pushnotifications.reporting.api.OpenEvent
 class OpenNotificationActivity: Activity() {
     private val log = Logger.get(this::class)
 
-    private fun startIntent(bundle: Bundle, clickAction: String? = null) {
+    private fun startIntent(bundle: Bundle?, clickAction: String? = null) {
         val i: Intent
         if (clickAction != null) {
             i = Intent()
             i.action = clickAction
         } else {
-            i = packageManager.getLaunchIntentForPackage(packageName)
+            i = packageManager.getLaunchIntentForPackage(packageName)!!
         }
 
         i.replaceExtras(bundle)
@@ -69,17 +68,17 @@ class OpenNotificationActivity: Activity() {
                    timestampSecs = System.currentTimeMillis() / 1000
                 )
 
-                val dispatcher = FirebaseJobDispatcher(GooglePlayDriver(applicationContext))
-                val job = dispatcher.newJobBuilder()
-                        .setService(ReportingJobService::class.java)
-                        .setTag("pusher.open.publishId=${pusherData.publishId}")
-                        .setConstraints(Constraint.ON_ANY_NETWORK)
-                        .setRetryStrategy(RetryStrategy.DEFAULT_EXPONENTIAL)
-                        .setLifetime(Lifetime.UNTIL_NEXT_BOOT)
-                        .setExtras(ReportingJobService.toBundle(reportEvent))
+                val reportWorker = OneTimeWorkRequest.Builder(ReportingWorker::class.java)
+                        .setConstraints(Constraints.Builder()
+                                .setRequiredNetworkType(NetworkType.CONNECTED)
+                                .build())
+                        .setInputData(ReportingWorker.toInputData(reportEvent))
                         .build()
 
-                dispatcher.mustSchedule(job)
+                val workManagerInstance = WorkManager.getInstance(applicationContext)
+                workManagerInstance.enqueueUniqueWork("pusher.open.publishId=${pusherData.publishId}",
+                        ExistingWorkPolicy.KEEP,
+                        reportWorker)
 
                 startIntent(intent.extras, pusherData.clickAction)
             } catch (_: JsonSyntaxException) {