+++ CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.9.0] - 2022-01-04
+
+### Fixed
+- Fix issue with calling stop() then calling start() again
+
+## [1.8.0] - 2021-11-11
+
+### Fixed
+- Fix issue #126 and #127
+
+## [1.7.0] - 2021-06-07
+
+### Fixed
+- Fix issue #112 and #117
+
+### Changed
+- Migrate to AndroidX
+- Update Firebase dependencies
+- Introduce Android WorkManager in place of Firebase JobDispatcher
+
 ## [1.6.2] - 2020-01-27
 
 ### Fixed